+++ client/src/components/Design/components/Amount/EditableAmount.tsx
@@ -4,7 +4,7 @@ import {useEffect, useRef, useState} from 'react';
 import {useTheme} from '@components/Design/theme/HDesignProvider';
 import TYPOGRAPHY from '@components/Design/token/typography';
 
-import Icon from '../Icon/Icon';
+import Icon from '../Icons/Icon';
 import Text from '../Text/Text';
 
 interface Props {
