+++ client/src/components/Design/components/Icons/IconSvg.tsx
