+++ client/src/components/Design/components/Icons/Icons/IconCheck.tsx
