+++ client/src/components/Design/components/Icons/Icons/IconConfirmCircle.tsx
