+++ client/src/components/Design/components/Icons/Icons/IconPictureSquare.tsx
