+++ client/src/components/Design/components/Icons/Icons/IconSearch.tsx
