+++ client/src/components/Design/components/Icons/Icons/IconTrash.tsx
