+++ client/src/components/Design/components/Icons/Icons/IconXCircle.tsx
