+++ client/src/components/Design/components/Icons/Svg.tsx
