+++ client/src/components/Design/components/Input/Input.tsx
@@ -3,7 +3,7 @@ import React, {forwardRef, useEffect, useImperativeHandle, useRef, useState} fro
 
 import IconButton from '@HDcomponents/IconButton/IconButton';
 import {InputProps} from '@HDcomponents/Input/Input.type';
-import Icon from '@components/Design/components/Icon/Icon';
+import Icon from '@components/Design/components/Icons/Icon';
 import {useTheme} from '@theme/HDesignProvider';
 
 import Flex from '../Flex/Flex';
