+++ client/src/components/Design/components/ListButton/ListButton.tsx
@@ -4,7 +4,7 @@ import React, {forwardRef} from 'react';
 import Text from '@HDcomponents/Text/Text';
 import IconButton from '@HDcomponents/IconButton/IconButton';
 import Flex from '@HDcomponents/Flex/Flex';
-import Icon from '@components/Design/components/Icon/Icon';
+import Icon from '@components/Design/components/Icons/Icon';
 import {useTheme} from '@theme/HDesignProvider';
 
 import {ListButtonProps} from './ListButton.type';
