+++ client/src/components/Design/components/Top/EditableLine.tsx
@@ -5,7 +5,7 @@ import {useEffect, useRef, useState} from 'react';
 import TYPOGRAPHY from '@components/Design/token/typography';
 import {useTheme} from '@components/Design/theme/HDesignProvider';
 
-import Icon from '../Icon/Icon';
+import Icon from '../Icons/Icon';
 
 interface Props {
   value: string;
